feat(Keyboard): Reorganize some classes
- "VRTK_GridKeyboardLayoutCalculator"
  - ->"VRTK_SimpleKeyboardLayoutCalculator"
- "VRTK_KeyboardRenderer"
  - -> "VRTK_BaseKeyLayoutRenderer"
  - -> "VRTK_CanvasKeyLayoutRenderer"
