+++ Assets/VRTK/Scripts/Keyboard/VRTK_BaseKeyLayoutRenderer.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3d2081c52a1d85c4f8a3bbe74ef44b25
+timeCreated: 1486103240
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
